prepare for custom resolver
move current Resolver into FutureResolver
make more private methods to mroe easily adapt the behavior
